Major reorganization
Hooks are now separated off into their own modules.
This freed up considerable space in the main.rs file
and makes it easier to navigate around.
