+++ Cargo.toml
@@ -10,13 +10,13 @@ gluster = "~0.7"
 init-daemon = "*"
 ipnetwork = "~0.12"
 itertools = "*"
-juju = "~0.5"
+juju = "~0.6"
 libc = "*"
 libudev = "*"
-nix = "*"
+nix = "~0.8"
 pnet = "~0.16"
 rand = "~0.3"
-regex = "*"
+regex = "~0.1"
 resolve = "~0.1"
 rustc-serialize = "*"
 serde = "^0.9"
