+++ src/hooks/brick_detached.rs
@@ -1,0 +1,4 @@
+pub fn brick_detached() -> Result<(), String> {
+    // TODO: Do nothing for now
+    return Ok(());
+}
