+++ src/hooks/mod.rs
@@ -1,0 +1,7 @@
+pub mod brick_attached;
+pub mod brick_detached;
+pub mod config_changed;
+pub mod fuse_relation_joined;
+pub mod nfs_relation_joined;
+pub mod server_changed;
+pub mod server_removed;
